chore(persons): add CI safety check to persons migrations
